create models in testing_system app
