+++ EDUPlatform/testing_system/admin.py
@@ -1,3 +1,10 @@
 from django.contrib import admin
 
-# Register your models here.
+from .models import Answer, Article, Course, Question, Test, Topic
+
+admin.site.register(Course)
+admin.site.register(Topic)
+admin.site.register(Article)
+admin.site.register(Test)
+admin.site.register(Question)
+admin.site.register(Answer)
