+++ EDUPlatform/users/admin.py
@@ -1,23 +1,7 @@
 from django.contrib import admin
 
-from .models import (
-    Answer,
-    Article,
-    Course,
-    Question,
-    Student,
-    Teacher,
-    Test,
-    Topic,
-    User,
-)
+from .models import Student, Teacher, User
 
 admin.site.register(User)
 admin.site.register(Teacher)
 admin.site.register(Student)
-admin.site.register(Course)
-admin.site.register(Topic)
-admin.site.register(Article)
-admin.site.register(Test)
-admin.site.register(Question)
-admin.site.register(Answer)
